+++ src/FileEntrySecret.cs
@@ -159,6 +159,15 @@ namespace CSCommonSecrets
 			return fileEntryAsDictionary;
 		}
 
+		/// <summary>
+		/// Get key identifier
+		/// </summary>
+		/// <returns>Key identifier as string</returns>
+		public string GetKeyIdentifier()
+		{
+			return System.Text.Encoding.UTF8.GetString(this.keyIdentifier);
+		}
+
 		#endregion // Common getters