+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 0.9.3 (not released yet)
+- Add GetKeyIdentifier to FileEntrySecret (**FEATURE**)
 
 ## Version 0.9.2 (released 2019-12-26)
 - Use correct version number with CommonSecretsContainer constructors (**FIX**)